Refactor imports to use operational modules
- Updated import paths in various files to point to the new operational modules for connection, query, index, collection management, and server functionalities.
- Adjusted test files to reflect the changes in import paths, ensuring that all tests continue to function correctly with the new structure.
- Consolidated imports for clarity and maintainability across the codebase.
