+++ src/cb_mcp/tools/operational/kv.py
@@ -17,10 +17,10 @@ import couchbase.subdocument as subdoc
 from couchbase.exceptions import CouchbaseException
 from fastmcp import Context
 
-from ..utils.connection import connect_to_bucket, format_keyspace
-from ..utils.constants import LOGGER_ROOT
-from ..utils.context import get_cluster_connection
-from ..utils.responses import tool_error, tool_success
+from ...utils.constants import LOGGER_ROOT
+from ...utils.context import get_cluster_connection
+from ...utils.operational.connection import connect_to_bucket, format_keyspace
+from ...utils.responses import tool_error, tool_success
 
 logger = logging.getLogger(f"{LOGGER_ROOT}.tools.kv")
