+++ src/cb_mcp/tools/operational/query.py
@@ -12,10 +12,10 @@ from fastmcp import Context
 from fastmcp.server.dependencies import get_access_token
 from lark_sqlpp import modifies_data, modifies_structure, parse_sqlpp
 
-from ..utils.connection import connect_to_bucket, format_keyspace
-from ..utils.constants import LOGGER_ROOT, SCOPE_WRITE
-from ..utils.context import get_cluster_connection
-from ..utils.query_utils import (
+from ...utils.constants import LOGGER_ROOT, SCOPE_WRITE
+from ...utils.context import get_cluster_connection
+from ...utils.operational.connection import connect_to_bucket, format_keyspace
+from ...utils.operational.query_utils import (
     evaluate_query_plan,
     extract_plan_from_explain_results,
 )
