+++ src/cb_mcp/utils/operational/certs/capella_root_ca.pem
