+++ src/cb_mcp/utils/operational/index_utils.py
@@ -13,7 +13,7 @@ from urllib.parse import urlparse
 
 import httpx
 
-from .constants import LOGGER_ROOT
+from ..constants import LOGGER_ROOT
 
 logger = logging.getLogger(f"{LOGGER_ROOT}.utils.index_utils")
 
@@ -264,14 +264,15 @@ def _get_capella_root_ca_path() -> str:
     """
     try:
         # Use importlib.resources to get the certificate path (works for installed packages)
-        cert_file = files("cb_mcp.certs").joinpath("capella_root_ca.pem")
+        cert_file = files("cb_mcp.utils.operational.certs").joinpath(
+            "capella_root_ca.pem"
+        )
         # Convert to string path - this works for both installed packages and dev mode
         return str(cert_file)
     except (ImportError, FileNotFoundError, TypeError):
-        # Fallback for development: use src/certs/ directory
-        utils_dir = os.path.dirname(os.path.abspath(__file__))
-        src_dir = os.path.dirname(utils_dir)
-        fallback_path = os.path.join(src_dir, "certs", "capella_root_ca.pem")
+        # Fallback for development: certs/ sits next to this module.
+        here = os.path.dirname(os.path.abspath(__file__))
+        fallback_path = os.path.join(here, "certs", "capella_root_ca.pem")
 
         if os.path.exists(fallback_path):
             logger.info(f"Using fallback certificate path: {fallback_path}")
