+++ src/providers/static.py
@@ -6,8 +6,8 @@ from typing import Any
 from couchbase.cluster import Cluster
 from fastmcp import Context
 
-from cb_mcp.utils.connection import connect_to_couchbase_cluster
 from cb_mcp.utils.constants import LOGGER_ROOT
+from cb_mcp.utils.operational.connection import connect_to_couchbase_cluster
 
 logger = logging.getLogger(f"{LOGGER_ROOT}.providers.static")
