+++ tests/unit/test_logger_names.py
@@ -34,17 +34,17 @@ EXPECTED_LOGGER_NAMES = {
     "cb_mcp.auth": "couchbase.auth",
     "cb_mcp.core.app": "couchbase.core.app",
     "cb_mcp.tool_registration": "couchbase.tool_registration",
-    "cb_mcp.tools.collection_management": "couchbase.tools.collection_management",
-    "cb_mcp.tools.index": "couchbase.tools.index",
-    "cb_mcp.tools.kv": "couchbase.tools.kv",
-    "cb_mcp.tools.query": "couchbase.tools.query",
-    "cb_mcp.tools.server": "couchbase.tools.server",
+    "cb_mcp.tools.operational.collection_management": "couchbase.tools.collection_management",
+    "cb_mcp.tools.operational.index": "couchbase.tools.index",
+    "cb_mcp.tools.operational.kv": "couchbase.tools.kv",
+    "cb_mcp.tools.operational.query": "couchbase.tools.query",
+    "cb_mcp.tools.operational.server": "couchbase.tools.server",
     "cb_mcp.utils.cli": "couchbase.utils.cli",
     "cb_mcp.utils.config": "couchbase.utils.config",
-    "cb_mcp.utils.connection": "couchbase.utils.connection",
+    "cb_mcp.utils.operational.connection": "couchbase.utils.connection",
     "cb_mcp.utils.elicitation": "couchbase.utils.elicitation",
     "cb_mcp.utils.environment": "couchbase.utils.environment",
-    "cb_mcp.utils.index_utils": "couchbase.utils.index_utils",
+    "cb_mcp.utils.operational.index_utils": "couchbase.utils.index_utils",
     "cb_mcp.utils.scope_enforcement": "couchbase.utils.scope_enforcement",
     "cb_mcp.utils.telemetry": "couchbase.utils.telemetry",
     "providers.static": "couchbase.providers.static",
