+++ tests/unit/test_query_plan_evaluation.py
@@ -1,7 +1,7 @@
 """Unit tests for query plan evaluation helpers."""
 
-from cb_mcp.tools.query import evaluate_query_plan
-from cb_mcp.utils.query_utils import extract_plan_from_explain_results
+from cb_mcp.tools.operational.query import evaluate_query_plan
+from cb_mcp.utils.operational.query_utils import extract_plan_from_explain_results
 
 
 def test_evaluate_query_plan_detects_primary_scan() -> None:
